Screenspace Raytracing
This is still an early build for the SSRTGI, I have not made any performance optimizations so expect a huge (or none) impact to performance compared to v1.1.

High preset has indirect lighting set to  SSRTGI by default, you may change it to SSRT in the shader settings.

SSRTGI is still a work in progress as well so it might not look right in some cases. There is also no proper filtering so expect an increase in grain in lighting.

- Removed SSRAO, replaced with SSRTAO
- Replaced GI with SSRTGI from Bliss Shaders
- Atmospheric Tweaks
- Lighting changes
- Added Subsurface Scattering to other blocks (i.e. sand, ice, cactus, and more)
- Tweaked shader presets
- Subsurface scattering properly applied to profiles High or higher
- Added Very High for SSRT

And more  :)