v2.0
v2.0 has received a massive overhaul on it's overall look, feel and atmospherics. Even though this shader is fairly young and hasn't received much updates, I began to feel as if the appearance was not going in a direction that I had originally intended. So I've spent the past couple of months revisioning what Allium should be, and what should have been. Although it's not perfect, I have been fairly happy with how it has turned out and wanted to share it with you all. Cheers! :)

Changelog:
- Overhauled Volumetric Clouds and it's lighting
- Temporarily removed Cloud Layers
- Crepuscular Rays
- Cloud Shadow
- New Sky
- Lighting changes
- Changed from ACES Tonemap to Lottes
- Bloom tweaks
- Applied SSS to new blocks such as sand, snow, etc (WIP)
- Removed Global Illumination and SSRAO
- Switched back to Complementary's SSAO
- Gas giant and space apearance in the End dimension
- Adjusted atmospherics for Distant Horizons